refactor: move Query updater to hook
